SmartSched_AI/smartsched/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
loguru==0.7.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
